MAINT: move CellResponse to separate file
