+++ hnn_core/__init__.py
@@ -1,7 +1,8 @@
 from .dipole import simulate_dipole, read_dipole, average_dipoles
 from .drives import drive_event_times
 from .params import Params, read_params
-from .network import Network, CellResponse, read_spikes
+from .network import Network, read_spikes
+from .cell_response import CellResponse
 from .pyramidal import L2Pyr, L5Pyr
 from .basket import L2Basket, L5Basket
 from .parallel_backends import MPIBackend, JoblibBackend
